Moved TimeoutException definition from timer module to auxiliary to prevent problem with importing multiprocessing
