+++ pilot/api/data.py
@@ -40,7 +40,7 @@ from pilot.util.filehandling import (
 from pilot.util.math import convert_mb_to_b
 from pilot.util.parameters import get_maximum_input_sizes
 from pilot.util.workernode import get_local_disk_space
-from pilot.util.timer import TimeoutException
+from pilot.util.auxiliary import TimeoutException
 from pilot.util.tracereport import TraceReport
